requirements.txt
Django==5.1.5
python-dotenv==1.1.1
razorpay==1.4.2
Requests==2.32.5
gunicorn
dj-database-url
python-decouple
psycopg2-binary 